moving toffoli_gate test to test_quantumprogram
